@@ -4,12 +4,39 @@
   "private": true,
   "description": "",
   "scripts": {
-    "dev-client": "webpack-dev-server",
+    "dev-client": "better-npm-run webpack-dev-client",
     "dev-server": "npm run server -- development",
-    "webpack": "webpack",
-    "server": "./node_modules/.bin/babel-node --presets react,es2015,stage-0 ./server/server.js",
-    "buildSchema": "./node_modules/.bin/babel-node --presets react,es2015,stage-0 ./buildSchema.js",
-    "start": "npm run buildSchema &&  npm run webpack && npm run server"
+    "server-babel": "./node_modules/.bin/babel-node  ./server/server.js",
+    "schema": "./node_modules/.bin/babel-node ./schema.js ",
+    "server": "node ./server/webpack/server.js",
+    "webpack-server": "better-npm-run webpack-server",
+    "webpack-client-dev": "better-npm-run webpack-client-dev",
+    "start": "npm run webpack-client-dev && npm run server-babel",
+    "start-server": "npm run webpack-server && npm run server"
+  },
+  "betterScripts": {
+    "webpack-dev-client": {
+      "command": "webpack-dev-server",
+      "env": {
+        "NODE_ENV": "development",
+        "SERVER": false
+      }
+    },
+    "webpack-server": {
+      "command": "webpack",
+      "env": {
+        "NODE_ENV": "development",
+        "SERVER": true,
+        "NODE_PATH": "./node_modules/"
+      }
+    },
+    "webpack-client-dev": {
+      "command": "webpack",
+      "env": {
+        "NODE_ENV": "development",
+        "SERVER": false
+      }
+    }
   },
   "engines": {
     "node": "4.2.x"
@@ -21,29 +48,33 @@
     "jquery": "^2.2.1",
     "ramda": "^0.21.0",
     "lodash": "^4.6.1",
-    "react": "^0.14.3",
-    "react-dom": "^0.14.3",
+    "react": "^15.2.1",
+    "react-dom": "^15.2.1",
     "react-link-state": "^0.1.2",
     "react-relay": "^0.9.0",
     "react-router": "^2.4.1",
-    "react-router-relay": "^0.13.2"
-    , "autobind-decorator": "^1.3.3",
-    "react-styleable": "^2.2.4"
+    "react-router-relay": "^0.13.2",
+    "autobind-decorator": "^1.3.3",
+    "react-styleable": "^2.2.4",
+    "isomorphic-relay": "^0.7.0",
+    "isomorphic-relay-router": "^0.8.0"
   },
   "devDependencies": {
+    "babel-plugin-css-modules-transform" :"^0.1.1",
+    "webpack-merge": "^0.14.0",
+    "better-npm-run": "^0.0.9",
+    "null-loader": "^0.1.1",
+    "extract-text-webpack-plugin": "^1.0.1",
+    "isomorphic-style-loader": "^1.0.0",
     "kerberos": "0.0.19",
-    "clean-webpack-plugin" :"^0.1.9",
     "mongodb": "^2.1.18",
     "cors": "^2.7.1",
     "express": "^4.13.3",
     "express-graphql": "^0.5.3",
-
     "graphql": "^0.6.0",
     "graphql-relay": "^0.4.2",
-
     "webpack": "^1.12.1",
     "webpack-dev-server": "^1.12.0",
-    "webpack-merge": "^0.2.0",
     "html-webpack-plugin": "^2.12.0",
     "json-loader": "^0.5.4",
     "react-hot-loader": "^1.3.0",
@@ -67,7 +98,7 @@
     "connect-flash": "^0.1.1",
     "body-parser": "^1.15.2",
     "cookie-parser": "^1.4.3",
-    "ejs": "^2.4.2",
+    "ejs": "^2.3.4",
     "classnames": "^2.2.5",
     "css-loader": "^0.23.1",
     "postcss": "^5.0.21",
@@ -83,3 +114,4 @@
     "postcss-selector-not": "^2.0.0"
   }
 }
+
