@@ -6,25 +6,18 @@ var precss = require('precss');
 var autoprefixer = require('autoprefixer');
 var postcssImport = require('postcss-import');
 var pkg = require('./package.json');
-var CleanWebpackPlugin = require('clean-webpack-plugin');
-var TARGET = process.env.npm_lifecycle_event;
+import merge from 'webpack-merge'
 
-const DEV_MODE = TARGET === 'dev';
+var ExtractTextPlugin = require('extract-text-webpack-plugin');
 
-console.warn("DEV_MODE: " + DEV_MODE);
-const deps = Object.keys(pkg.dependencies);
-module.exports = {
+const SERVER = process.env.SERVER === 'true';
+const DEVELOPMENT = process.env.NODE_ENV === "development";
 
-  entry:{
-    vendor: deps,
-    main: path.resolve(path.resolve(__dirname), './js/app.js')
-  }
- ,
-  output: {
-    path: path.resolve(__dirname, './server/static'),
-    chunkFilename: '/[id].chunk.js',
-    filename: DEV_MODE ? '[name].[chunkhash].js' : '[name].[chunkhash].min.js'
-  },
+console.log("DEVELOPMENT " + DEVELOPMENT)
+console.log("SERVER " + SERVER)
+console.log("NODE_PATH "+process.env.NODE_PATH)
+
+const common = {
 
   postcss: function (webpack) {
     return [precss, autoprefixer
@@ -38,11 +31,11 @@ module.exports = {
     loaders: [
       {
         test: /\.js$/, exclude: /node_modules/,
-
         loader: "babel",
         query: {
+          // ignore babelrc, which is used for some package.json scripts.
+          babelrc: !!SERVER,
           cacheDirectory: true,
-          // plugins: ['transform-decorators-legacy'],
           plugins: ['transform-decorators-legacy', 'transform-runtime', '../../../babelRelayPlugin'],
           presets: ['es2015', 'stage-0', 'react'],
         }
@@ -51,23 +44,13 @@ module.exports = {
 
       , {
         test: /\.css$/,
-        // loaders: ['style', 'css?modules&localIdentName=[local]---[hash:base64:5]', 'cssnext'],  //  2st - css, 1 - css
-        // loaders: ['style', 'css?modules&localIdentName=[local]---[hash:base64:5]', 'cssnext'],  //  2st - css, 1 - css
-        loader: "style-loader!css-loader?modules&localIdentName=[local]---[hash:base64:5]&importLoaders=1!postcss-loader",
+        // loader: ExtractTextPlugin.extract("isomorphic-style-loader", "css-loader?modules&localIdentName=[local]---[hash:base64:5]&importLoaders=1", "postcss-loader"),
+        loader: ExtractTextPlugin.extract("style-loader", "css-loader?modules&localIdentName=[local]---[hash:base64:5]&importLoaders=1", "postcss-loader"),
         exclude: /node_modules/
       }
-
-
+      , {test: /\.md$/, loader: 'null'}
+      , {test: /^inputfile[0-9]+$ /, loader: 'null'}
       , {test: /\.json$/, loader: 'json-loader'}
-
-      // ,   {
-      //   test: /\.html$/,
-      //   loader: "html",
-      //   include: [
-      //     path.resolve(__dirname, "/public"),
-      //
-      //   ],
-      // }
 
     ]
   }
@@ -75,45 +58,107 @@ module.exports = {
   ,
   devServer: {
     historyApiFallback: true,
-    // hot: true,
     inline: true,
     progress: true
     , port: 5000
   }
 
+}
 
-  , node: {
-    fs: 'empty',
-  }
+if (SERVER) {
+  console.log("IN SERVER CONFIG")
+  module.exports = merge(common, {
+    target: "node",
+    node: {
+      __filename: true,
+      __dirname: true,
+      console: true
+    },
+    entry: {
+      server: path.resolve(__dirname, "./server/server.js"),
+      // main: path.resolve(__dirname, './src/client.js')
+    },
+    output: {
+      path: path.resolve(__dirname, './server/webpack'),
+      filename: 'server.js',
+      libraryTarget: 'commonjs2'
+    }
+
+    , plugins: [
+
+      new webpack.DefinePlugin({
+        __DEV__: DEVELOPMENT,
+        'process.env.NODE_ENV': JSON.stringify(DEVELOPMENT ? 'development' : 'production')
+      }),
+      // new ExtractTextPlugin('styles.[chunkhash].css'),
+      new ExtractTextPlugin('styles.css'),
+      // new ExtractTextPlugin("[name]-[chunkhash].css"),
+
+      // new webpack.optimize.CommonsChunkPlugin({
+      //   name: 'vendor',
+      //   filename: 'vendor.[hash].js',
+      //   minChunks: function (module, count) {
+      //     return module.resource && module.resource.indexOf(path.join(__dirname, '..', 'node_modules')) === 0;
+      //   }
+      // }),
+
+    ]
+  })
+} else {
+  console.log("IN CLIENT CONFIG")
+  const deps = Object.keys(pkg.dependencies);
+
+  module.exports = merge(common, {
+    entry: "./src/client.js",
+    node: {fs: 'empty'},
+    // output: {
+    //   path: path.resolve(__dirname, './server/static'),
+    //   chunkFilename: '/[id].chunk.js',
+    //   // filename: DEVELOPMENT ? '[name].[chunkhash].js' : '[name].[chunkhash].min.js'
+    //   filename: DEVELOPMENT ? '[name].[chunkhash].js' : '[name].[chunkhash].min.js'
+    // },
+    //  SPLIT
+    //  entry:{
+    //    vendor: deps,
+    //    app: path.resolve(__dirname, './src/client.js')
+    //  } ,
+    output: {
+      path: path.resolve(__dirname, './server/static'),
+      chunkFilename: '/[id].chunk.js',
+      // filename: DEVELOPMENT ? '[name].js' : '[name].min.js'  // app filename not vendor
+      filename: DEVELOPMENT ? 'app.js' : 'app.min.js'  // app filename not vendor
+    },
+     // output: {
+     //   path: path.resolve(__dirname, './server/static'),
+     //   chunkFilename: '/[id].chunk.js',
+     //   filename: DEVELOPMENT ? '[name].[chunkhash].js' : '[name].[chunkhash].min.js'
+     // },
+    plugins: [
+      new webpack.HotModuleReplacementPlugin(),
+      new ExtractTextPlugin('styles.css'),
+      new webpack.DefinePlugin({
+        __DEV__: DEVELOPMENT,
+        'process.env.NODE_ENV': JSON.stringify(DEVELOPMENT ? 'development' : 'production')
+      }),
+
+      // new ExtractTextPlugin('styles.[chunkhash].css'),
+
+      // new ExtractTextPlugin('[name].css'),
+      // new ExtractTextPlugin("[name]-[chunkhash].css"),
+
+      // new webpack.optimize.CommonsChunkPlugin('vendor', 'vendor.js')
+
+      // new webpack.optimize.CommonsChunkPlugin({
+      //   name: 'vendor',
+      //   // filename: 'vendor.[hash].js',
+      //   filename: 'vendor.js',
+      //   minChunks: function (module, count) {
+      //     return module.resource && module.resource.indexOf(path.join(__dirname, '..', 'node_modules')) === 0;
+      //   }
+      // })
+    ]
+  })
+}
 
 
-  , plugins: [
-    new webpack.HotModuleReplacementPlugin(),
-    new CleanWebpackPlugin(['static'], {
-      root:  path.resolve(path.resolve(__dirname), './server'),
-      verbose: true,
-      dry: false
-    })  ,
-    new webpack.DefinePlugin({
-      __DEV__: DEV_MODE,
-      'process.env.NODE_ENV': JSON.stringify(DEV_MODE ? 'development' : 'production')
-    }),
-    // new webpack.DefinePlugin({"global.GENTLY": false}),
-    new webpack.optimize.CommonsChunkPlugin({
-      name: 'vendor',
-      filename: 'vendor.[hash].js',
-      //minChunks: 2
 
-
-      minChunks: function (module, count) {
-        return module.resource && module.resource.indexOf(path.join(__dirname, '..', 'node_modules')) === 0;
-      }
-    }),
-
-    new HtmlwebpackPlugin({
-      title: 'relay_app'
-    })
-  ]
-
-
-};
