@@ -1,6 +1,7 @@
 import express from 'express';
 import cors from 'cors';
-
+import fs from 'fs'
+import {MongoClient}  from 'mongodb';
 import UserSchema from '../data/userschema'
 import GraphQLHTTP from 'express-graphql'
 import path from 'path';
@@ -9,19 +10,23 @@ import cookieParser from 'cookie-parser';
 import flash from 'connect-flash';
 import passport from 'passport';
 import session from 'express-session'
-
-import {MongoClient}  from 'mongodb';
+import ejs from 'ejs';
+import   {graphql}   from 'graphql';
+import {introspectionQuery} from 'graphql/utilities'
 import setUpPassport from './setUpPassport';
 
 import routes from './routes';
 import database  from "../data/database";
+import renderOnServer from './renderOnServer'
 let app = express();
 app.use(cors());
 //  schema development easing
-process.devmode = process.argv.indexOf('development') != -1;
-console.warn("process.devmode " + process.devmode)
-var staticPath = path.join(__dirname, "static");
-app.use(express.static(staticPath));
+//todo find a better way to set a global variable
+process.devmode = process.env.NODE_ENV === "development";
+console.warn("process.devmode " + process.devmode);
+
+// const staticPath = path.join(__dirname, "static");
+// app.use(express.static(staticPath));
 
 app.set("views", path.join(__dirname, "views"));
 app.set("view engine", "ejs");
@@ -30,12 +35,14 @@ app.set("view engine", "ejs");
 (async() => {
   try {
     //  hardcoded pass
-    let db = await MongoClient.connect("mongodb://root:1234@ds015849.mlab.com:15849/rgrs");
-
+    let db = await MongoClient.connect("mongodb://root:1234@ds015849.mlab.com:15849/rgrs")
     let schema = UserSchema(db);
 
-    const dbManager = database(db);
+    // Serve CSS
+    app.use('/css/', express.static(path.resolve(__dirname, './static')));
 
+
+    const dbManager = database(db);
     setUpPassport(dbManager)();
     app.use(/\/((?!graphql).)*/, bodyParser.urlencoded({extended: true}));
     app.use(/\/((?!graphql).)*/, bodyParser.json());
@@ -48,10 +55,9 @@ app.set("view engine", "ejs");
     app.use(flash());
     app.use(passport.initialize());
     app.use(passport.session());
+    //
     app.use(routes);
-
     app.use('/graphql', GraphQLHTTP(request => {
-
           return ({
             schema,
             context: request.session,
@@ -61,9 +67,55 @@ app.set("view engine", "ejs");
               locations: error.locations,
               stack: error.stack
             })
+            , pretty: true
           })
         })
     )
+
+    // Serve JavaScript
+    app.get('/app.js', (req, res) => {
+      res.setHeader('Content-Type', 'application/javascript');
+      res.sendFile('app.js', {root: __dirname + '/static/'});
+    });
+    //
+    app.get('/vendor.js', (req, res) => {
+      res.setHeader('Content-Type', 'application/javascript');
+      res.sendFile('vendor.js', {root: __dirname + '/static/'});
+    });
+
+    // Serve HTML
+    app.get('/*', (req, res, next) => {
+      renderOnServer(req, res, next);
+    });
+
+    /// catch 404 and forwarding to error handler
+    app.use(function (req, res, next) {
+      const err = new Error('Not Found');
+      err.status = 404;
+      next(err);
+    });
+
+
+//     development error handler
+// will print stacktrace
+    app.use(function (err, req, res, next) {
+      res.status(err.status || 500);
+      res.send({
+        message: err.message,
+        error: err
+      });
+    });
+
+
+    // production error handler
+    // no stacktraces leaked to user
+    // app.use(function (err, req, res, next) {
+    //   res.status(err.status || 500);
+    //   res.send({
+    //     message: err.message,
+    //     error: {}
+    //   });
+    // });
 
     app.set("port", process.env.PORT || 3000);
 
@@ -71,11 +123,13 @@ app.set("view engine", "ejs");
       console.log("listening on port " + app.get("port"))
     });
 
+
+
   } catch (e) {
     console.log(e)
   }
 
 
-})()
+}) ()
 
 
